refactor: place schedule struct in its own file
