+++ contracts/non-fungible-tokens/andromeda-marketplace/src/testing/tests.rs
@@ -14,7 +14,7 @@ use andromeda_std::{
         },
         encode_binary,
         expiration::{Expiry, MILLISECONDS_TO_NANOSECONDS_RATIO},
-        Milliseconds, Schedule,
+        Milliseconds, schedule::Schedule,
     },
     error::ContractError,
     testing::mock_querier::MOCK_CW20_CONTRACT,
